+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## [7.0.0]
+### Changed
+- Updated to [Sophia 7.0.0](https://github.com/aeternity/aesophia/blob/master/CHANGELOG.md#700---2021-10-20)
+### Removed
+- Support for AEVM
+- Backend parameters
+- `decode-data` endpoint (it was used by AEVM exclusively)
+
 ## [6.1.0]
 ### Changed
 - Updated to [Sophia 6.1.0](https://github.com/aeternity/aesophia/blob/master/CHANGELOG.md#610---2021-10-20)
@@ -124,7 +132,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial HTTP interface: /aci, /compile, /decode-data, /encode-calldata, /version, /api-version, /api
 - Docker support (aeternity/aesophia_http)
 
-[Unreleased]: https://github.com/aeternity/aesophia_http/compare/v6.1.0...HEAD
+[Unreleased]: https://github.com/aeternity/aesophia_http/compare/v7.0.0...HEAD
+[7.0.0]: https://github.com/aeternity/aesophia_http/compare/v6.1.0...v7.0.0
 [6.1.0]: https://github.com/aeternity/aesophia_http/compare/v6.0.2...v6.1.0
 [6.0.2]: https://github.com/aeternity/aesophia_http/compare/v6.0.1...v6.0.2
 [6.0.1]: https://github.com/aeternity/aesophia_http/compare/v6.0.0...v6.0.1